Problem 96.
Doesn't do any deduction beyond the simple first-level constraints.
If this deduction fails, brute forces a square with the lowest number
of possible solutions.  Still runs in about 20 seconds in ruby.

Prints using the Ruby benchmark library to see which ones are most
difficult to solve in this manner.

A good future task would be do come up with additional layers of
deduction to see if more can be solved without brute force.