Ada pr014: Use sufficiently large integer type
Instead of relying on the 'Natural' or 'Long_Natural' types being long
enough, specify a type that is specifically known to be long enough.

On x86-64, Long_Natural is a 64-bit type, but on at least 32-bit ARM,
it is only a 32-bit type (probably x86 as well).  Use a type that will
work consistently.