COBOL: much faster problem 4 solution
Instead of using multiplication and division to reverse the digits of
the number, take advantage of 'usage display' which represents the
numbers as a string of digits.  Just reverse the digits textually.

This results in about a 10-fold performance increase.  It seems that
cobc uses bignums for intermediate results whenever multiplication or
division is involved.