Haskell 82-83
Problems 81-83 are all variants on the same graph traversal problem,
and are easiest to solve with the same code.

To solve the other variants, Change the 'weights' mapping from an
array to a general mapping.  This allows adding a special "exit" node
for problem 82.  It doesn't seem to have a significant impact on the
performance, which isn't surprising, since most of the time is
probably spent sorting the nodes by sums, and rebuilding the node tree
with the updated values.