Ada: Clean up the prime sieve
Move the Sieve object into a child package inside of the prime sieve,
and only allocate it internally.  It needs to be in a package so that
the object system will work (even though we don't use it, it still
generates a warning).  Next_Prime can now properly allocate the sieve
as it needs to.