rust: Update for rust 2014-12-25 snapshot
There have been lots of minor changes to the language.  The generic
numbers have been simplified a bit.  Most of the rest are moving
deprecated calls to the curren ones, and some cleanups for things that
are more strict now.

Most notable, indexing now works with other things, so that foo[n] and
foo[n] = value; both now work on Vecs.