rust: Autogenerate problem list
Use a cargo build hook to generate the problem list before running the
compilation.

Unfortunately, this isn't perfect.  The attribute that specifies a
module path only works with literals.  The include!() macro built-in
to rust doesn't allow 'mod' or 'use' declarations.  We work around
this by creating a symlink src/plist.rs that links back to the
generated file.