rust: Allow solutions to be checked
Extend the `define_problem!()` macro to accept a third argument, which
is the expected solution to the problem.  When defined with 3
arguments, the function called should return something of a type
compatible with this third argument, and the test invocation will use
`assert_eq!()` to ensure the test returns the right result.

In addition, problems registered with 3 arguments will register a test
function so that the solution will be checked when all tests are built
and run.