rust: Play with permutation iterators
Rust's libstd contains a permutation iterator that could be useful for
some of these problems.  Since it clones the result each time, as this
benchmark code shows, it is quite a bit slower (about 6.5 times for
me).  Aside from performance, it returns the permutation in a
single-swap order instead of in lexicographical order.  The swap order
is easier to compute, but some of the Euler problems need the
lexicographical.

This change also leaves in several attempts to make an iterator that
can avoid copying.  None of these can pass the borrow checker.  It
seems to be an incompatibility between the iterator protocol, and
something that wants to return the whole container.  Since the
iterator must mutably hold the data, it isn't possible to (safely) get
at this same underlying data.

All in all, this code doesn't actually change any of the permutation
code, since, clumsy as it is to use, it is the most efficient way to
get all of the permutations.