rust: Clean up build process
Instead of wedging a symlink for 'src/plist.rs' in the build, just
have this file include relative to the OUT_DIR environment variable.
To make this work, we need to put absolute paths to the source files
for each 'mod' statement.  This overcomes the strong desire rust has
to always reference files relative to the source file referencing
them.