haskell: Bring in 10
There is some experiments in this to do with different prime seive
types.  Benchmarking indicates that the STUArray-based one is by far the
fastest, taking about 223ms for primes under 2 million (compare with the
27ms for the Rust solution).

The somewhat naive Haskell solution with a lazy array and a isPrime
function that references back to the array building it is about 4x
slower.

The complicated lazy sieve with the 'next' values kept for each prime is
faster for smaller values, but slows down quite a bit as we approach 2
million.