Chez scheme 1-24
These should be mostly r6rs.  There are some references to the srfi
libraries, which are available at https://github.com/arcfide/chez-srfi,
the `link-dirs.chezscheme.sps` script should be run so that Chez scheme
can find the files.  It is easies to check this out as a subdirectory of
the chez directory, as "." will already be in the search path.