lisp: Start conversion to a single program
Instead of individual files that have to be manually compiled and loaded
in the right order, begin the process of converting the lisp solutions
to be a single package.  Create a bit of driver code, and a main script
(using Roswell) to be able to run from the command line.  Currently, the
main program just runs all registered problems, in numeric order.

Problems now register themselves, and problems 1-25 have been converted.

This also captures the beginning of changing some of the problems to
start using LOOP instead of ITER.  ITER is more powerful, so some
constructs can't be represented with loop, and multiple-variable usage
is less efficient with LOOP, so some might be left with ITER.