swift: Move all problems into structs/classes
Instead of just top level functions, and just a comment with the correct
answer, move the problems into structs or classes that implement
the `Problem` protocol, which can verify that each problem gives the
correct result.

As part of this, fix problem 6 to actually give the correct result, by
changing the loop bounds to match the problem description.